[refactor] useObserver hook으로 리팩토링
