+++ src/api/posts.js
@@ -1,6 +1,6 @@
 import http from '@/api/http';
 
-export const get = (limit = 10, page = 1) =>
+export const get = (limit = 20, page = 1) =>
   http.get(`/posts?sortBy=createdAt:desc&limit=${limit}&page=${page}`);
 
 export const insert = data => http.post(`/posts`, data);
